refactor: Move app navigation components to appNavigation module
This commit moves several app-level navigation components from the `app` module to the `appNavigation` module. This is done to better encapsulate navigation logic and improve modularity.

The following files have been moved:

- `AflamiNavBar.kt`
- `AppNavGraph.kt`
- `AppNavigatorImpl.kt`
- `AppScaffold.kt`
- `ObserveAsEvents.kt`

The following changes were also made:

- Updated `app/src/main/java/com/paris_2/aflami/di/FeaturesApiModule.kt` to import `AppNavigatorImpl` from its new location.
- Updated `app/src/main/java/com/paris_2/aflami/MainActivity.kt` to use `AppScaffold` from the `appNavigation` module.
- Added necessary dependencies to `appNavigation/build.gradle.kts` including design system and various feature APIs.
