+++ app/src/main/java/com/paris_2/aflami/MainActivity.kt
@@ -4,6 +4,7 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
+import com.paris_2.aflami.appnavigation.AppScaffold
 import com.paris_2.aflami.designsystem.theme.AflamiTheme
 
 class MainActivity : ComponentActivity() {
